[Refactor] 패키지 변경
- extension 패키지를 생성하여 관련 기능들을 이동
- extension 컨트롤러에 매핑할 url 주소를 변경
