+++ src/main/java/flow/assignment/extension/model/request/DeleteCustomExtensionReq.java
@@ -1,9 +1,9 @@
-package flow.assignment.file.model.request;
+package flow.assignment.extension.model.request;
 
 import lombok.Getter;
 
 @Getter
-public class PostExtensionCheckedReq {
+public class DeleteCustomExtensionReq {
     private String extension;
     private String option;
 }
