+++ src/main/java/flow/assignment/interceptor/ExtensionHandlerInterceptor.java
@@ -5,8 +5,8 @@ import flow.assignment.common.filetype.GeneralFile;
 import flow.assignment.common.filetype.MimeType;
 import flow.assignment.common.error.ErrorCode;
 import flow.assignment.common.error.exception.InterceptorException;
-import flow.assignment.file.model.entity.Extension;
-import flow.assignment.file.repository.ExtensionRepository;
+import flow.assignment.extension.model.entity.Extension;
+import flow.assignment.extension.repository.ExtensionRepository;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.servlet.http.Part;
